LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/origin-keyed-agent-clusters/popups/opener-yes-openee-yes-same.sub.https-expected.txt


PASS message event must occur
PASS setting document.domain must give sync access
FAIL opener: originAgentCluster must equal true assert_equals: expected (boolean) true but got (undefined) undefined
FAIL openee: originAgentCluster must equal true assert_equals: expected (boolean) true but got (undefined) undefined
